crashtesting: fix PDF export of fdo83057-2.docx
Loading this document and laying it out resulted in an assertion failure
since commit de90c192cb8f1f03a4028493d8bfe9a127a76b2a (sw content
controls, plain text: enable DOCX filter with data binding, 2022-09-19).

Writer has two relevant invariants: First, if a draw shape is anchored
into a header, then the "master" shape (typically on the first page) and
its "virtual" shapes (on other pages) have their indexes / ZOrders set
so that the master shape has the highest index. The second is that in
case a draw+fly format is paired into a "textbox", then the fly index is
always the draw index + 1. The assert fails because in this case the
virtual draw shape's index is not smaller than the master one.

What seems to happen here is that first SwDrawContact::ConnectToLayout()
iterates the frames of the text node in the header, and happens to first
visit page 1 and then page 2. So the master SdrObjects go to the first
page, the virtual ones go to the second page. Then later
SwDrawContact::ConnectToLayout() is called again, which removes the
master and the virtual draw SdrObject from the layout and the virtual
draw SdrObject from the model. Then it visits the text frames of the
header text node again, but this time it happens to find the second
page, and only then the first page (for reasons unclear to me, this
didn't happen before the above commit). The result is that both
invariants are failing: the master draw+fly is no longer on the same
page and the indexes are also out of order, so one of the pages don't
show shape text for the shape from the header.

Fix the problem by two tweaks:

- Improve SwDrawContact::ConnectToLayout(), so that in case it just
  appends the draw object to the layout for the master, then it adjusts
  its ZOrder so the matching fly will be the draw one + 1.

  This was working in the virtual case already, since there we call
  AddVirtObj() already, which calls SwDrawVirtObj::AddToDrawingPage(),
  which already knows how to maintain this invariant.

  This fixes master draw SdrObjects to be directly before their fly
  counterparts, even if the page order changes.

- Improve SwDrawVirtObj::AddToDrawingPage(), so that in case the virtual
  draw SdrObject finds its fly, but that fly has an incorrect index (due
  to the changed page visit order), then we first fix up the fly and
  only then set the index of the virtual draw SdrObject.

The result is a doc model like this:

<SdrObject ptr="0x52ecd30" symbol="13SwDrawVirtObj" name="" title="" description="" nOrdNum="0" aOutRect="-16000, -16000, 0, 0"/>
<SwVirtFlyDrawObj ptr="0x4e4bfa0" fly-frame="5">
  <SdrObject ptr="0x4e4bfa0" symbol="16SwVirtFlyDrawObj" name="" title="" description="" nOrdNum="1" aOutRect="0, 0, 0, 0"/>
<SdrObject ptr="0x4ea4300" symbol="17SdrObjCustomShape" name="Rectangle 2" title="" description="" nOrdNum="2" aOutRect="10418, 2441, 1775, 485">
<SwVirtFlyDrawObj ptr="0x52ec630" fly-frame="14">
  <SdrObject ptr="0x52ec630" symbol="16SwVirtFlyDrawObj" name="" title="" description="" nOrdNum="3" aOutRect="0, 0, 0, 0"/>

I.e. each draw SdrObject is followed by its fly and the masters have the
highest indexes.

This all stems from the behavior that 1) re-connecting a draw contact to
the layout removes the draw shapes but not the flys from the layout, 2)
the order of frames in an SwIterator visiting is not something we can
depend on and 3) textboxes are implemented using a pair of draw+fly
formats.

Change-Id: Ia03f0ec70cf67626660349accc63770fe2afef5d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140688
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins