tdf#72456: Don’t cache macOS font list
The “system” font list is used for both actual system fonts and fonts we
add with AddTempDevFont(), so the caching makes loading embedded fonts
broken the first time the file is opened, requiring a next open without
closing the application.

The callers, I think, are careful not to call GetDevFontList() unless
they really want to refresh the font list, so the caching is probably
superfluous if not wrong.

Change-Id: Ie14cc473a64425fd5270bc65ffd61dea2d873cfe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140840
Tested-by: Jenkins
Reviewed-by: خالد حسني <khaled@aliftype.com>