tdf#151387 Fix regression cause by tdf#149279
The gtk_tree_view_expand_to_path function not only expands all parent
rows of path as necessary but also expands the children of the row of
the path. This explains the difference seen between gtk3inst and
salinst when a collapsed row is scrolled to. gtk3inst expands the
collapsed row, salinst does not. The enhancement patch for tdf#149279
removed the gtk_tree_view_expand_to_path function from
gtk_tree_view_scroll_to_row. This caused a regression in the styles
tree. To fix the regression this patch reverts the removed
gtk_tree_view_expand_to_path functions. To make the enhancement patch
behave the same for gtk3 and sal, the scroll to row is collapsed
after scrolling if that was it's state before.

Change-Id: I3c3975a3f258c6c432eb866a1c712299e2faf5be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141048
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>