tdf#124098: sc, ods import: do not recalculate row heights
if we have already optimal heights in the xml, because it is
not necessary and it takes a lot of time loading a file.

Citing here for completeness:
a) if there is no height attribute and the "use_optimal-row-height"
attribute is set then calculate the row height

b) if there is a height attribute just use it and do not recalculate.

Maybe TODO for later: optimize row height calculation more in sc

Change-Id: I9d964aad744970e5d36f239c0ce39ce98c00f273
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141204
Tested-by: Jenkins
Tested-by: Balazs Varga <balazs.varga.extern@allotropia.de>
Reviewed-by: Balazs Varga <balazs.varga.extern@allotropia.de>