tdf#151548 sw: use provided name for formfields
This was already aded for TextFieldmark in 2018 via tdf#120225.
(A unique name is created if that mark already exists.)

Added it to the other formfield items.
This is needed for DOC/DOCX import,
and is critical for connecting macros to the control.

mstahl said that  DateFieldmark is not a valid MS formfield,
and miklosv indicated  no likely value in preserving a name there.

There is also NonTextFieldmark and FieldmarkWithDropDownButton,
but they just look like implementation gadgets to me
so I didn't give them the same treatment.

No need for a unit test here. This is super foundational - anything
else building on it will be the unit test.

Change-Id: Ide49d6efb8391fea17e7a61c9e99b30532bb2014
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141423
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>