tdf#105609: Support DateModified for updated tdoc stream contents
...so that the Python script provider, based on those DateModified values, will
reload a script embedded in a document after that script has (programmatically)
been modified.

As long as a stream content in a tdoc document has not been modified, it will
report a default-initialized (all zero) DateModified.  Only when the stream has
been modified will that be changed to the current date.  While that might not be
the most beautiful implementation, it at least gets the job done of fixing
tdf#105609 "Python script provider does not reload modified embedded scripts".

(The DateModified values cannot be stored directly in the tdoc_ucp::Content
instances, as those are thrown away and recreated on demand.  So they needed to
be stored more persistently at the tdoc_ucp::OfficeDocumentsManager.)

Change-Id: Iee809960e1a1bc40961f0df2b3729e58b75e6028
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141491
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>