documentLoadWithOptions takes URLs
m_aDocPath is set in lok_doc_view_set_property from data
coming from gtv_application_open's calls to g_file_get_path.
So the path is not a proper file URL, which is expected by
documentLoadWithOptions. Some Linux paths can happen to be
able to convert to absolute URL in rtl::Uri::convertRelToAbs,
but not all, and not e.g. Windows paths.

Change-Id: I0a17cb84000603336a6809b0179bbda3d54aec23
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141579
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>