sc lok: Double-click should behave more like on desktop
When editing was introduced in Calc LibreOfficeKit, there were lots of
limitations.  Particularly the thinking was that it would be good if a
double-click into any text (even if spanning over more cells)
actually started editing that text.

These days, the LOK behaves much more consistently, so it is better to
behave as the LibreOffice on desktop (or other tools, like gdocs) do:

If the cell:

* Is empty

    -> single click places the cell cursor
    -> double click places the cell cursor & shows the text caret

* Is empty, but covered by text bleeding from other cell

    -> same as if it was empty

* Isn't empty

    -> single click places the cell cursor
    -> double click places the cell cursor & places the text caret
       inside the text where the user clicked

        + this is actually different in gdocs - there the caret is
	  placed at the end of the text; in LO it is where the user has
	  double-clicked

Change-Id: Ib5884f887c98f803b06d8bed5057ec435be480ef
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142196
Tested-by: Jenkins
Reviewed-by: Jan Holesovsky <kendy@collabora.com>