tdf#149189 detect and handle graphic change on Sw OLE
If in any way a used SW OLE Chart gets changed
(example in task uses a script and UNO API), detect
this and react accordingly. Use getGraphicVersion()
from EmbeddedObjectRef to decide when to reset bufferd
graphic data. It gets incremented at all occasions
where the graphic changes due to being used for a
similar case already.
An alternative would be to extend SwOLEListener_Impl
with a XModifyListener as it is done in
EmbedEventListener_Impl, that would require all the
(add|remove)ModifyListener calls and/ managing these,
plus having a 2nd listener to these when
EmbeddedObjectRef already provides that. Tried that
this works also if an alternative would be needed.

Change-Id: Ic27088b0241aa4b5ecfc3c2b6e64994d720b3546
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141803
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>