Consistently set document view also for zoom event
And extract this code to a dedicated function.

The problem was, that creating a new view sets the document's current
view shell (in SwViewGlueDocShell); upon destruction, document's view
shell is set to nullptr. In desktop case, document view shell is then
reset to another active shell e.g. inside "grab focus" event, which
makes all following operations safe. But in gtktiledviewer case, the
"grab focus" event is not handled, so setClientZoomInThread resulted
in eventual nullptr dereference.

It was workarounded in commit 25de85bc3ce2d2f9b7736f65492f42579048da27
  Date   Fri Oct 21 21:14:04 2022 +0200
    nullptr check

but that was not the proper fix.

Possibly it could also be reasonable to start handling "focus-in-event"
[1] in gtktiledviewer; but since just setting the view consistently in
setClientZoomInThread solves the problem, it looks excessive.

This reverts commit 25de85bc3ce2d2f9b7736f65492f42579048da27.

[1] https://docs.gtk.org/gtk3/signal.Widget.focus-in-event.html

Change-Id: Iac4ff4e9b043aec0e0af6ed811c0e7f018378271
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141753
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>