Do not merge filesarrayref and unixlinksarrayref when filtering out directories
In commit 8c6467a785037309d2a2cec56c8e38cf52da0aee, the code
filtering out the directories was creating a new array, that
was assigned to the function's inout filesarrayref argument;
and that code merged non-directories from both filesarrayref
and unixlinksarrayref.

This fixes the logic to only filter filesarrayref.

Change-Id: Ia00dcd3f645ba2a6c2d8ea0559627ccfe0eb232b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142195
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>