SdNavigator: Make the page objects tree stand still when the entry
selected is already visible (affects tdf#138868 and tdf#129610)

This patch replaces the use of TreeView::scroll_to_row with
TreeView::set_cursor in SdPageObjsTLV::SelectEntry because the GtkInst
scroll_to_row always scrolls the entry to the top, when possible, even
if the entry is already in the visible entries range. SalInst only
scrolls the entry to the top when it is not in the visible range.
Commit 3c8e81669f2311dd35aa6787cbf3e5a6a0520433 'tdf#149279 related:
make gtk treeview scroll_to_row more sal like' made the gtk behavior
similiar to the sal behavior.

Change-Id: I95aa3555699eb8c8d88bb00ca39d7c3a6589353a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142363
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>