tdf#148719 Moving caption frame & undoing is shrinking caption frame
So the problem here is that when we call SwDoc::FindFormatByName we can
have the situation that there are two formats with the same name.

Those formats appear to have which ids
   159   RES_DRAWFRMFMT
   156   RES_FLYFRMFMT

Previously, the index was sorted by
   ... which ... name

which meant that the lower-valued which was always been returned.

After my path, the index is sorted by
   ...name....which

The problem is call path is
   sw/source/core/undo/unattr.cxx:252 in SwUndoFormatAttr::GetFormat
where it calls
   rDoc.GetSpzFrameFormats()->FindFormatByName()

To fix it, use the findByTypeAndName method to explicitly specify which
which-id we are interested in.

Change-Id: I8941f7c277785afb1940bcf5d2b5153c291b6f97
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142371
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>