sw content controls: drop not needed XServiceInfo impl
As noted in
<https://gerrit.libreoffice.org/c/core/+/142454/3#message-18fe8bb36fc24d317ad16dd25e236ed51c88a910>,
introducing a dedicated css.text.ContentControls service is actually not
needed and client code should be fine with just an "anonymous" object
returned by getContentControls() + implementing XIndexAccess.

Remove it before somebody starts to depend on it. Let's rather have a
bit of inconsistency (e.g. SwXFootnotes implements XServiceInfo, while
SwXContentControls not) than an XServiceInfo implementation that we'll
have to support from now on, without a user.

Change-Id: Ifde4fb6bcafdffabb189447415b89b01c9675296
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142511
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins